Always show line numbers by default
When submitting a paper to SSTIC, a reviewer commented that the line
numbers where missing from the code listings. As it is a good idea to
include them, enable them by default in `sstic.cls`.

If an author wants to include a listing with no line numbers, it is
possible to use `numbers=none` directly in the `\begin{lstlisting}`
parameters. This is also documented in `sstic.cls`.